+++ src/ResourceModel.ts
@@ -1,6 +1,40 @@
+import { ResourceCRUD } from './ResourceCommon/ResourceCRUD';
 import { ResourceHelper } from './ResourceHelper';
 
 export abstract class ResourceModel {
+
+  static resourceInstance: ResourceCRUD<any, any, any> = null;
+
+  static get(id: string): Promise<any> {
+    return this.getInstance().get({id});
+  }
+
+  static query(query?: any): Promise<any[]> {
+    return this.getInstance().query(query);
+  }
+
+  static remove(id: string): Promise<void> {
+    return this.getInstance().remove({id});
+  }
+
+  private static getInstance(): ResourceCRUD<any, any, any> {
+    if (!this.resourceInstance) {
+
+      const model: ResourceModel = (new (this as any)());
+
+      if (!model.$resource) {
+        throw new Error('Your resource is not defined');
+      }
+
+      if (!model.$resource.instance) {
+        throw new Error('Your resource is not created (inject it somewhere)');
+      }
+
+      this.resourceInstance = (new (this as any)()).$resource.instance;
+    }
+
+    return this.resourceInstance;
+  }
 
   abstract readonly $resource: any = null;